added option to exclude sees that look werid
I also made some major changes to the species complexes. I added a
bunch of species and collapsed non-sect. peltigera species complexes
down to the section. There are two species complexes that I haven’t
added data for, so they need to be commented out or the whole pipeline
fails